ScrolledListener removed to separated class
